+++ composer.json
@@ -3,7 +3,7 @@
     "type": "template",
     "require": {
         "doctrine/annotations": "^1.13",
-        "mamluk/kipchak": "0.33",
+        "mamluk/kipchak": "0.46",
         "symfony/yaml": "^6.2",
         "zircote/swagger-php": "^4.11"
     },
